Move pytest_addoption call to top-level conftest.py
It only works there, see https://github.com/pytest-dev/pytest/issues/1427
