+++ lib/matplotlib/tests/conftest.py
@@ -2,6 +2,6 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 from matplotlib.testing.conftest import (
-    mpl_test_settings, pytest_addoption, pytest_configure, pytest_unconfigure,
+    mpl_test_settings, pytest_configure, pytest_unconfigure,
     pytest_sessionfinish, pytest_terminal_summary)
